+++ joplin/base/models/information_page.py
@@ -1,12 +1,17 @@
 from django.db import models
 
+from modelcluster.fields import ParentalKey
+from modelcluster.models import ClusterableModel
+
 from wagtail.core.fields import RichTextField, StreamField
 from wagtail.core.blocks import RichTextBlock
-from wagtail.admin.edit_handlers import FieldPanel, InlinePanel, StreamFieldPanel
+from wagtail.admin.edit_handlers import FieldPanel, InlinePanel, MultiFieldPanel, PageChooserPanel, StreamFieldPanel
+from wagtail.snippets.edit_handlers import SnippetChooserPanel
 
 from base.forms import InformationPageForm
 
 from .janis_page import JanisPage
+from .contact import Contact
 
 from .constants import WYSIWYG_GENERAL
 
@@ -48,3 +53,43 @@ class InformationPage(JanisPage):
         FieldPanel('additional_content'),
         InlinePanel('contacts', label='Contacts'),
     ]
+
+class InformationPageRelatedDepartments(ClusterableModel):
+    page = ParentalKey(InformationPage, related_name='related_departments', default=None)
+    related_department = models.ForeignKey(
+        "base.departmentPage",
+        on_delete=models.PROTECT,
+    )
+
+    panels = [
+        # Use a SnippetChooserPanel because blog.BlogAuthor is registered as a snippet
+        PageChooserPanel("related_department"),
+    ]
+
+class InformationPageContact(ClusterableModel):
+    page = ParentalKey(InformationPage, related_name='contacts')
+    contact = models.ForeignKey(Contact, related_name='+', on_delete=models.CASCADE)
+
+    panels = [
+        SnippetChooserPanel('contact'),
+    ]
+
+    def __str__(self):
+        return self.contact.name
+
+class InformationPageTopic(ClusterableModel):
+    page = ParentalKey(InformationPage, related_name='topics')
+    topic = models.ForeignKey('base.TopicPage',  verbose_name='Select a Topic', related_name='+', on_delete=models.CASCADE)
+    toplink = models.BooleanField(default=False, verbose_name='Make this page a top link for this topic')
+
+    panels = [
+        MultiFieldPanel(
+            [
+                PageChooserPanel('topic'),
+                FieldPanel('toplink'),
+            ]
+        ),
+    ]
+
+    def __str__(self):
+        return self.topic.text
